Enhance Eventix application with new features and fixes
This commit introduces several key improvements to the Eventix application, including:
- Updated SQL queries for better readability in `GetAllCategoriesHandler.cs` and `GetAllTicketTypesHandler.cs`.
- Added validation rules in various validator classes to ensure command integrity.
- Introduced new command and handler classes for archiving and creating categories, as well as for ticket types.
- Created new response models to standardize API responses.
- Added new endpoint mappings for CRUD operations on categories and ticket types.
- Improved error handling in the domain layer for clearer failure messages.
- Refactored existing classes for better structure and maintainability.

+++ src/Modules/Events/Eventix.Modules.Events.Domain/TicketTypes/Entities/TicketType.cs
@@ -6,7 +6,7 @@ namespace Eventix.Modules.Events.Domain.TicketTypes.Entities
 {
     public sealed class TicketType : Entity
     {
-        private TicketType(Guid eventId, string name, decimal price, string currency, int quantity)
+        private TicketType(Guid eventId, string name, decimal price, string currency, decimal quantity)
         {
             EventId = eventId;
             Specification = (name, quantity);
@@ -20,7 +20,7 @@ namespace Eventix.Modules.Events.Domain.TicketTypes.Entities
         public TicketTypeSpecification Specification { get; private set; } = null!;
         public Money Price { get; private set; } = null!;
 
-        public static TicketType Create(Guid eventId, string name, decimal price, string currency, int quantity)
+        public static TicketType Create(Guid eventId, string name, decimal price, string currency, decimal quantity)
             => new(eventId, name, price, currency, quantity);
 
         public void UpdatePrice(decimal price)
