+++ src/Modules/Events/Eventix.Modules.Events.Domain/TicketTypes/ValueObjects/TicketTypeSpecification.cs
@@ -2,16 +2,16 @@
 {
     public record TicketTypeSpecification
     {
-        public TicketTypeSpecification(string name, int quantity)
+        public TicketTypeSpecification(string name, decimal quantity)
         {
             Name = name;
             Quantity = quantity;
         }
 
         public string Name { get; }
-        public int Quantity { get; }
+        public decimal Quantity { get; }
 
-        public static implicit operator TicketTypeSpecification((string name, int quantity) value)
+        public static implicit operator TicketTypeSpecification((string name, decimal quantity) value)
             => new(value.name, value.quantity);
 
         public override string ToString() => $"{Name} ({Quantity})";
